Update CandidatePoolApplications rather than re-creating them
We used to delete all CandidatePoolApplications records and create them
again.

This is fine but eventually we'll hit a limit of how many ids we can
generate.

To fix this. This commit adds some SQL UPSERT logic. It relies on an
unique index on the application_form_id.

If there is a conflict we update the record.
If there is no conflict on the application_form_id we create a new
record.
If there CandidatePoolApplications that should be deleted we delete
them.

This will replicate the existing behaviour without recreating records
every 15 minutes