Lb/1351 evaluate possibly remove detect invariant daily checks (#11261)
## Context

We run jobs called DetectInvariantChecks hourly and daily. It sends
alerts to Sentry when an ‘Invariant’ is detected. This is a real hodge
podge of stuff and we tend to ignore it. I did a bit of investigation
(see the trello ticket) about what jobs were worth keeping / moving
elsewhere / deleting

## Changes proposed in this pull request

See the commit history for description of different changes

## Guidance to review


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency