Fix pool opt in confirmation page
This page can be technically accessed at any time. Even when candidates
don't have a published preference.

Not through the UI directly but through back links or if the candidate
knows the url.

A small number of users to this and this commit redirect the user to the
application_sharing page if the state of the preference is not correct.