[lb/1385] Select main site where available (#11278)
## Context

When a provider has said they do not want schools to be selectable by
the candidate, then we want to select the 'Main site' if that is
available, rather than just selecting the first available.

## Changes proposed in this pull request

- Method on the site model for determining if it is a main site
- Setting the option with a main site if available, otherwise using the
first.

## Guidance to review

On the review app, sign in as a candidate and create an application
choice for UCL (does not allow candidates to choose school)
Sign in as a support user and view the application choice -- the main
site should be chose.

On the review app, sign in as a candidate and create an application
choice for GORSE SCITT (_does_ allow candidates to choose school).
Sign in as a support user and view the application choice -- it should
be the site you chose as a candidate.


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency