Select main site for courses with study modes (#11283)
There's an edge case we are trying to fix where the provider doesn't allow candidates to choose the site and the course they select has multiple study modes.

In this scenario we should select the main site if it exists. Otherwise select the first available site.

Previously we would just select the first site rather than the main site.