[lb/1394] Use grade instead of band score on review component (#11289)
## Context

[Sentry
error](https://dfe-teacher-services.sentry.io/issues/6936720646/?alert_rule_id=853774&alert_type=issue&notification_uuid=48c23077-a037-444f-a1aa-8762ff0d03fb&project=1765973&referrer=slack)

This was taking place at the point of submission when reviewing the
application. The candidate entered the name for the
OtherEflQualification `IELTS` (it is free text).

## Changes proposed in this pull request

There is a `grade` method on all of the EFL qualification models that
delegates to the appropriate method. So rather than using `band_score`
we can just use grade and it will use the correct method for the kind of
EFL qualification that was selected.

## Guidance to review

<!-- How could someone else check this work? Which parts do you want
more feedback on? -->


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency