[lb/1395] Reverts random wait time that ensured notify rate limit not reached (#11292)
## Context

This was a temporary measure put in place for the first week to ensure
we didn't hit the notify rate limit during a surge of applications. Now
things are back to normal levels of activity, we can revert this and
queue up emails immediately.

## Changes proposed in this pull request

Reverting changes from earlier PR. 

## Guidance to review



## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency