Migrate duplicated preferences to new cycle (#11293)
When candidates with pool preferences with opt_in status carry-over we create a duplicated preference for them, attached to the new cycle application form.

The issue here is that these duplicated preferences don't become published, and because the preferences are scoped to the candidate. A candidate in this scenario will not be prompted to confirm their duplicated preferences because the system thinks that the candidate has confirmed(published) their preferences.

So these candidates will probably never confirm their preferences. So we will do that for them.