Bug: Candidate receives change_course email if the only change is location (#11304)
## Context

Bug 🐞: The candidate receives an email describing changes to the course
in cases where the site has changed but this change is not notifiable
given that the choice does not have placement autoselect enabled, i.e.
they've not selected or seen the site to need to know of any changes.

This was discovered as part of another bug 🐛: 'The application choice's
"Location" is defaulting to a Placement School that isn't main site'.

## Changes proposed in this pull request

- Added checks for course change
- This includes if the site is changed AND the school has placements
autoselected enabled
- This prevents emails from being sent when site is changed but school
does not have this setting
- Added specs for service

## Guidance to review

- Please double-check the logic in the service and the spec scenarios.
My understanding from the existing code is that
`application_choice.school_placement_auto_selected` makes site updates
meaningful and otherwise any changes to site are not notifiable changes,
but this may be more complex than I've interpreted.
- If there are ways to simulate receiving an email within local, let me
know and I do more out-of-spec testing.

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency