Add missing attributes for TTA adviser sign up (#11308)
## Context

The TTA team got in touch as they would like us to start sending some
additional data along with the quickfire signups so they can monitor
this better -- apparently some of the dashboards were broken because the
data hasn't been included.

## Changes proposed in this pull request

Additional three fields added to the object we send to TTA sign up API

## Guidance to review

I've tested this with the TTA team and they are happy that the correct
data is being sent.


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency