Clarity implementation (#11274)
## Context

We want to be able to capture how people use our service outside of
interviews.

## Changes proposed in this pull request

- Fix a bug where GA cookies weren't actually being cleared
- JS script which responds to cookie consent to add clarity script and
cookies / clear cookies
- Feature flag to allow the feature to be turned on and off via the
support console.
- Content on the cookie's page (my words, so not good ones)


## Guidance to review
This is being implemented for both providers and candidates, so you can
do any of this testing as either a provider or a candidate.

### Testing cookies
Scenario one: Do not consent to cookies
navigate to any page in the service
open up the console in your browser and run 
```
clarity('metadata', (d, upgrade, consent) => {
       console.log('consentStatus:', consent);
  }, false, true, true);
```

It should raise an error as clarity should not exist on the page

Examine your cookies. You should not see the MUID cookie. And `_clck`
and `_clsk`, should be empty strings.

Scenario two: Consent to cookies
navigate to any page in the service
open up the console in your browser and run 
```
clarity('metadata', (d, upgrade, consent) => {
       console.log('consentStatus:', consent);
  }, false, true, true);
```

it should return a consentStatus object that looks like
```
{ad_Storage: 'denied', analytics_Storage: 'granted'}
```

Examine your cookies. You should see the MUID cookie. And `_clck` and
`_clsk` should have values.

### Testing recordings
- Accept cookies, and then do some stuff. 
- Wait a few minutes
- Sign into Clarity -- you should be able to find recordings for what
you've done
- Decline cookies. 
- Wait a few minutes
- Sign into Clarity -- you shouldn't see any of your new activity


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency