Lb/accessibility audit tickets 20 oct (#11313)
## Context

We need to finish a number of accessibility related tickets before we
have an external audit.
This PR is for 2 trello cards, but 4 are attached because we identified
some duplicates.

## Changes proposed in this pull request

- The address block in the Contact information section for candidates
should now read as a single block without the VO announcing 'Empty
group' between each line. I've fixed the issue in the SummaryComponent,
but also other places throughout where the `br` tag is used.
- Where the link to Find opens in a new tab for candidate viewing an
application, they will now hear and see the text 'opens in new tab'.

## Guidance to review

You can check this all using the mac VO, or any other screen reader type
utility.
On a mac, to use VO,
- Press Command+F5 to activate VoiceOver for Mac
- Use Control+Option+right arrow or left arrow to tab through each item
on the screen

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency