Allow providers to change course in pending condition (#11307)
We want to allow providers to change the course even when the application choice is in pending conditions. Right now this is done by the support dev.

This PR creates a new component that will show the change links on the offer page, similar to when the application choice is in the offer state. The provider can change the course details if the application choice is in offer state.

We reuse the change course form, with an injected hint for this action to notify the provider that they need to notify the candiate before making the change.

This means that the application_choice offer show controller needs to initialize different types of wizards depending on the state of the application_choice.