Create migration file for new column on DeferredOfferConfirmation (#11316)
## Context

This is a PR to migrate a new column to the `DeferredOfferConfirmation`
table called `offered_course_option` which will store the offer
(`application_choice.current_course_option`) before a deferred offer is
altered.

This new column will be used for testing as well as the email which
accompanies changes to confirm deferrals with new details.

## Changes proposed in this pull request

- Migration file which makes the above addition.

## Guidance to review

- Please review the migration file

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [x] This code adds a column or table to the database
  - [x] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency