Add temporary job to fix 2026 uuids in sandbox (#11328)
## Context

Because of a combination of cycle switching and testing, we have a bunch
of courses in sandbox manage that match codes, but not uuids in sandbox
publish. This breaks our incremental and full syncs.

We sorted the problem for UCL by just deleting all their 2026 courses
and application choices and and syncing from scratch.

We don't want to do that for all providers because they are actively
using the data.

## Changes proposed in this pull request

A Job which we'll remove once we are successfully syncing courses. It
only corrects the UUIDs.

After I have run this job, I'll run a full sync for sandbox and
everything should be up-to-date.

## Guidance to review

1. Login to the rails console on the review app or locally and change
the UUID of a 2026 course to something random (`SecureRandom.uuid`)
2. Run the job. 
3. The UUID of that course will change back to what it is in publish qa.


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency