Bug fix in HESA data export (#11341)
## Context

The `QLAIM` value in the `HesaDataExport` is incorrect for many courses.

The logic is currently based on the `Course` name starting with “QTS”
which many courses do not start with.

## Changes proposed in this pull request

Update the logic so that it returns:
- `020` if the qualifications array on the `Course` includes "pgce"
- `021` if the qualifications array on the `Course` includes "pgde"
- Leave blank if `undergraduate_degree` or anything other than the above

## Guidance to review

- Please review the spec to make sure that the `QLAIM` export row is
tested adequately.
- [Download the hesa
data](https://apply-review-11341.test.teacherservices.cloud/provider/reports/hesa-exports)
when the review app deploys and cross check the code in the `QLAIM` row
with the corresponding course either in the UI or in the console.

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency