Update git api faraday client gem
When we introduced the one login gem, the
get_into_teaching_api_client_faraday was conflicting with the one login
gem. Because of the different faraday versions the two gems were using.

To get around the issue. We created a PR in git.
https://github.com/DFE-Digital/get-into-teaching-api-ruby-client/pull/106

the version of the gem from this PR was used in apply to make the two
gems play nice with each other.

The git team now have merged that PR into main so we can use the main
version of the gem. Which this commit tries to do.