Add "apply filters" to application choices index (provider interface) (#11366)
## Context

When we were researching and designing Find candidates we found that
people did not like to scroll back to the top of a long list of filters
to apply them. They expected the apply button to be at the bottom, where
they are when they have finished adding all their filters.

We did this for the Find candidates filters but we should do it for the
other filters in Manage too.

## Changes proposed in this pull request

- Create a method in `provider_applications_filter` that activates the
show_bottom_button method in the filter_component source code:

_app/components/utility/filter_component.rb_
```
  def show_bottom_button?
    if @filter.respond_to?(:show_bottom_button?)
      @filter.show_bottom_button?
    else
      false
    end
  end
```

## Guidance to review

- Log in as a provider
- Visit the applications index
- Check that an 'Apply filters' button exists at the bottom of the
filter component and works as expected


https://github.com/user-attachments/assets/a3cab8ce-6998-4a10-b972-b9b1f2643cac

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency