Candidate interface link audit for new tab behaviour (#11367)
## Context

Currently, there isn’t a standard for when we open links in the same
window as opposed to opening in a new tab. Research has shown candidates
are sometimes nervous about clicking on links and losing their place in
the service.

This card is to audit links across the service and document where we
need to update targets of the links.

## Changes proposed in this pull request

- Add `new_tab: true` argument to `govuk_link` or
`govuk_link_to_with_utm_params` where the destination is outside of
Apply.

## Guidance to review

- Please visit the places in the review app where the argument has been
newly added and make sure it behaves as expected.

_Some instances:_


https://github.com/user-attachments/assets/2c5b02d0-f644-43ed-b709-627e94b4e21f


https://github.com/user-attachments/assets/903b313b-2d29-4300-828e-cc7c3f517acd


https://github.com/user-attachments/assets/38e72bca-385e-4e04-835e-6e58c6a6baeb

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency