Remove "Does the candidate have the potential to teach?" row key name (#11374)
## Context

When viewing a completed reference, the SummaryCard row containing the
feedback has the key "Reference". A previous design inserted logic so
that when the reference was carried over, the key was conditionally
rendered as "Does the candidate have the potential to teach?".

<img width="1875" height="1136" alt="Reference_process_map_-_Page_1_(3)"
src="https://github.com/user-attachments/assets/199b87aa-5ce7-4f35-9873-56590e44d884"
/>

This is not in line with current designs.

## Changes proposed in this pull request

- Remove the conditional row key text change.

## Guidance to review

- Log in as a provider
- View a candidate who has received feedback
- If their application form is not already carried over from a previous
cycle, manipulate the record so that reference.duplicate = true
- Check that the relevant key does not change 

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency