Add course_id to chasers_sent (#11378)
Some chasers are specific to the application choice and course. So this adds a column for the course_id.

This should fix the edge case where a chaser is sent to an application choice and the course on that application choice changes. It will allow us to send a chaser again.