Send reminder of visa sponsorship deadline approaching
We want to notify our candidates that have application choices in draft
state that the visa sponsorship deadline is approaching.

Currently that is 1 month prior to the deadline.

Because the chaser email is specific to the application choice and
course code. I've added a column to the chasers_sent table, to store the
course code.

This should address edge cases where candidates change application
choice to another visa sponsorship course. We would need to email them
again, and with chasers scoped to course code we would be able to do
that.