Bring the review page of make offer in line with other review pages (#11384)
When a provider makes an offer they end up on a review page that is not in line with our other review pages, mainly because the conditions pending and ske conditions are not part of the summary component.

This brings them in line with our other review pages. I could not use the standard Summary component we have for ske conditions so I just created one. To the user it looks like a big review list.