Fix activity audit logs page for inactive applications
When an application would become inactive after a set number of days, it
would be recorded on the activity page. But because the action is
automated and the ActivityLog::ApplicationChoice didn't add a specific
title to this event.

The page would not add any description to this event.

This commit should fix this.