Send reference email to referee that have not responded on accepted offer (#11390)
We send a reference request email to the referees of candidates when a candidate accepts an offer.

In the edge case when a candidate doesn't accept the offer and the referees have not responded yet. The next time the candidate accepts an offer we don't resent the reference request to the referee.

We want to send a reference request to all referees that have not responded every time the candidate accepts an offer.