Add uniq to location_preference strings on provider view (#11414)
## Context

A bug was identified during the Bug Party: “Find a candidate relaunch”
on 12 November 2025 whereby a provider could view the same location
preference and radius multiple times if it was duplicated by the
candidate.

We should only show the a location description once, it looks like a bug
even if it isn’t.

nb. We have discussed and decided to allow duplicate
location_preferences on the candidate side.

## Changes proposed in this pull request

- Add uniq to the list of strings mapped from candidate
location_preferences in the FAC provider view


https://github.com/user-attachments/assets/f62298ea-8d0a-4df1-b7f4-4ad467c6d3ee

## Guidance to review

- In your local make sure FAC is open and populated
- Log in as a candidate
- Update your sharing preferences and add multiple locations with the
same radius and location
- Log in as a provider and view the candidate's page in FAC
- Scroll to the bottom and check that the string pertaining to that
location only appears once

## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [ ] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency