lb/1490 Bug broken sentry integration (#11418)
## Context

Our sentry integration is broken because the `before_send` event in our
configuration converts the event to a hash. This was deprecated. See
sentry-rails release notes for 6.0 and this
[pr](https://github.com/getsentry/sentry-ruby/pull/2716/files)

## Changes proposed in this pull request

We will want to fix the `before_send` configuration as a long term fix,
but this will at least mean that we start seeing sentry events again.

## Guidance to review

<!-- How could someone else check this work? Which parts do you want
more feedback on? -->


## Things to check

- [ ] If the code removes any existing feature flags, a data migration
has also been added to delete the entry from the database
- [ ] API release notes have been updated if necessary
- [ ] If it adds a significant user-facing change, is it documented in
the [CHANGELOG](CHANGELOG.md)?
- [x] Attach the PR to the Trello card
- [ ] This code adds a column or table to the database
  - [ ] This code does not rely on migrations in the same Pull Request
- [ ] decide whether it needs to be in analytics yml file or analytics
blocklist
- [ ] data insights team has been informed of the change and have
updated the pipeline
- [ ] the sanitise.sql script and
0025-protecting-personal-data-in-production-dump.md ADR have been
updated
  - [ ] does the code safely backfill existing records for consistency